reordering configuration
