+++ Tutorials/redditclone/reddit_backend/src/main/java/com/JavaSpringPractice/ReditClone/config/SwaggerConfiguration.java
@@ -1,10 +1,6 @@
 package com.JavaSpringPractice.ReditClone.config;
 
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
@@ -14,10 +10,8 @@ import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
-@Configuration
 @EnableSwagger2
-@EnableWebMvc
-public class SwaggerConfiguration implements WebMvcConfigurer {
+public class SwaggerConfiguration{
     @Bean
     public Docket redditCloneApi(){
         return new Docket(DocumentationType.SWAGGER_2)
@@ -38,12 +32,4 @@ public class SwaggerConfiguration implements WebMvcConfigurer {
                 .build();
     }
 
-    @Override
-    public void addResourceHandlers (ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/swagger-ui.html**")
-                .addResourceLocations("classpath:/META-INF/resources/swagger-ui.html");
-        registry.
-                addResourceHandler("/webjars/**")
-                .addResourceLocations("classpath:/META-INF/resources/webjars/");
-    }
 }
